Move files into src folder
