+++ src/player.js
