+++ src/utils/winner-of-board.js
